Extract AdditionalDeductionsTooltip into its own file
Mirror the NetEmploymentIncomeTooltip pattern: the component now renders its own
DetailedTooltip with the per-tax title derived from taxType, so the two call
sites in TaxesTab no longer repeat the DetailedTooltip wrapper and title. The
ADDITIONAL_DEDUCTION_INFO and AdditionalDeductionsResult imports move with it.
