+++ src/components/TakeHomeCalculator/tabs/TaxesTab.tsx
@@ -15,19 +15,15 @@ import TableCell from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
-import type {
-  TakeHomeResults,
-  TakeHomeInputs,
-  AdditionalDeductionsResult,
-} from '../../../types/tax';
+import type { TakeHomeResults, TakeHomeInputs } from '../../../types/tax';
 import type { DependentDeductionResults } from '../../../types/dependents';
 import { formatJPY } from '../../../utils/formatters';
-import { ADDITIONAL_DEDUCTION_INFO } from '../additionalDeductionInfo';
 import AccountBalanceIcon from '@mui/icons-material/AccountBalance';
 import WarningIcon from '@mui/icons-material/Warning';
 import { DetailedTooltip } from '../../ui/Tooltips';
 import { ResultRow } from '../ResultRow';
 import NetEmploymentIncomeTooltip from './NetEmploymentIncomeTooltip';
+import AdditionalDeductionsTooltip from './AdditionalDeductionsTooltip';
 import { getNationalBasicDeductionTiers } from '../../../data/nationalBasicDeduction';
 
 interface TaxesTabProps {
@@ -205,87 +201,6 @@ const DependentDeductionTooltip: React.FC<DependentDeductionTooltipProps> = ({
           </>
         )}
       </Box>
-    </Box>
-  );
-};
-
-interface AdditionalDeductionsTooltipProps {
-  deductions: AdditionalDeductionsResult;
-  taxType: 'national' | 'residence';
-}
-
-const AdditionalDeductionsTooltip: React.FC<AdditionalDeductionsTooltipProps> = ({
-  deductions,
-  taxType,
-}) => {
-  const isNational = taxType === 'national';
-  const getAmount = (item: AdditionalDeductionsResult['items'][number]) =>
-    isNational ? item.national : item.residence;
-  const rows = deductions.items.filter(item => getAmount(item) > 0);
-  const total = isNational ? deductions.national : deductions.residence;
-
-  return (
-    <Box sx={{ minWidth: { xs: 0, sm: 320 }, maxWidth: { xs: '100vw', sm: 460 } }}>
-      <Typography variant="subtitle2" sx={{ fontWeight: 600, mb: 0.5 }}>
-        Additional Deductions Breakdown
-      </Typography>
-      <TableContainer component={Box} sx={{ mb: 2 }}>
-        <Table
-          size="small"
-          sx={{ '& .MuiTableCell-root': { padding: '2px 6px', fontSize: '0.95em' } }}
-        >
-          <TableHead>
-            <TableRow>
-              <TableCell>Deduction</TableCell>
-              <TableCell align="right">Amount</TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map(item => (
-              <TableRow key={item.key}>
-                <TableCell>
-                  <div style={{ fontWeight: 500 }}>
-                    {ADDITIONAL_DEDUCTION_INFO[item.key]?.name ?? item.label}
-                  </div>
-                </TableCell>
-                <TableCell align="right">{formatJPY(getAmount(item))}</TableCell>
-              </TableRow>
-            ))}
-            <TableRow sx={{ backgroundColor: 'action.hover' }}>
-              <TableCell sx={{ fontWeight: 600 }}>Total</TableCell>
-              <TableCell align="right" sx={{ fontWeight: 600 }}>
-                {formatJPY(total)}
-              </TableCell>
-            </TableRow>
-          </TableBody>
-        </Table>
-      </TableContainer>
-      <Typography variant="body2" sx={{ fontSize: '0.85em', color: 'text.secondary', mb: 1 }}>
-        {isNational
-          ? 'These reduce your taxable income for national income tax.'
-          : 'These reduce your taxable income for residence tax. The life and earthquake insurance deductions are smaller for residence tax than for income tax.'}
-      </Typography>
-      {rows.map(item => {
-        const info = ADDITIONAL_DEDUCTION_INFO[item.key];
-        if (!info) return null;
-        return (
-          <Typography
-            key={item.key}
-            variant="body2"
-            sx={{ fontSize: '0.85em', color: 'text.secondary', mt: 1 }}
-          >
-            <strong>{info.name}:</strong> {info.explanation}{' '}
-            <a
-              href={info.sourceUrl}
-              target="_blank"
-              rel="noopener noreferrer"
-              style={{ color: 'var(--primary-main)', textDecoration: 'underline' }}
-            >
-              {info.sourceLabel}
-            </a>
-          </Typography>
-        );
-      })}
     </Box>
   );
 };
@@ -649,12 +564,10 @@ const TaxesTab: React.FC<TaxesTabProps> = ({ results, inputs }) => {
             label={
               <span>
                 Other Deductions
-                <DetailedTooltip title="Other Income Deductions (National Tax)">
-                  <AdditionalDeductionsTooltip
-                    deductions={results.additionalDeductions}
-                    taxType="national"
-                  />
-                </DetailedTooltip>
+                <AdditionalDeductionsTooltip
+                  deductions={results.additionalDeductions}
+                  taxType="national"
+                />
               </span>
             }
             value={formatJPY(-results.additionalDeductions.national)}
@@ -1109,12 +1022,10 @@ const TaxesTab: React.FC<TaxesTabProps> = ({ results, inputs }) => {
             label={
               <span>
                 Other Deductions
-                <DetailedTooltip title="Other Income Deductions (Residence Tax)">
-                  <AdditionalDeductionsTooltip
-                    deductions={results.additionalDeductions}
-                    taxType="residence"
-                  />
-                </DetailedTooltip>
+                <AdditionalDeductionsTooltip
+                  deductions={results.additionalDeductions}
+                  taxType="residence"
+                />
               </span>
             }
             value={formatJPY(-results.additionalDeductions.residence)}
